refactor: structure mmoomm handlers
